Move the convert workflow to operations, for issue #372
