+++ src/mpfb/ui/operations/rigops/properties/keep_meta.json
@@ -1,0 +1,7 @@
+{
+    "type": "boolean",
+    "name": "keep_meta",
+    "description": "Do not delete the meta rig after producing. Only relevant when converting a MakeHuman-imported Game Engine rig",
+    "label": "Keep meta rig",
+    "default": false
+}
