+++ src/mpfb/ui/operations/rigops/properties/name.json
@@ -1,0 +1,7 @@
+{
+    "type": "string",
+    "name": "name",
+    "description": "Name to use when generating the final rig. If not specified, the name set in the panel \"object data properties\" -> \"advanced options\" -> \"rig name\" will be used. Only relevant for the MakeHuman-imported convert workflow",
+    "label": "Name",
+    "default": ""
+}
