+++ src/mpfb/ui/operations/rigops/properties/produce.json
@@ -1,0 +1,7 @@
+{
+    "type": "boolean",
+    "name": "produce",
+    "description": "After converting the rig to a rigify meta rig, also run the produce step to produce the final rig. Only relevant for the MakeHuman-imported convert workflow",
+    "label": "Produce",
+    "default": true
+}
