+++ src/mpfb/ui/rigging/__init__.py
@@ -4,7 +4,6 @@ _LOG.trace("initializing rigging module")
 
 from .rigpanel import MPFB_PT_Rig_Panel  # parent panel — must be first
 from .addrig import *
-from .rigify import *
 from .righelpers import *
 from .applypose import *
 from .addcycle import *
