--- src/mpfb/ui/rigging/rigify/operators/__init__.py
@@ -1,9 +1,0 @@
-from .....services import LogService
-_LOG = LogService.get_logger("rigifyoperators.init")
-_LOG.trace("initializing rigify operators")
-
-from .converttorigify import MPFB_OT_Convert_To_Rigify_Operator
-
-__all__ = [
-    "MPFB_OT_Convert_To_Rigify_Operator"
-    ]
