--- src/mpfb/ui/rigging/rigify/properties/keep_meta.json
@@ -1,7 +1,0 @@
-{
-    "type": "boolean",
-    "name": "keep_meta",
-    "description": "Do not delete the meta rig after producing",
-    "label": "Keep meta rig",
-    "default": false
-}
