--- src/mpfb/ui/rigging/rigify/properties/name.json
@@ -1,7 +1,0 @@
-{
-    "type": "string",
-    "name": "name",
-    "description": "Name to use when generating the final rig. If not specified, the name set in the panel \"object data properties\" -> \"advanced options\" -> \"rig name\" will be used",
-    "label": "Name",
-    "default": ""
-}
