+++ test/tests/eee_ui/operations/rigops/__init__.py
